split render role into PowerAssert.Renderer
